refactor: change name add-backlight-key module
